+++ src/pages/favorites/favorites.tsx
@@ -1,12 +1,14 @@
 import { Helmet } from 'react-helmet-async';
 
-import { useAppSelector } from '../../hooks';
-import { Favorites, Locations } from '../../components';
 import { CITIES } from '../../mocks';
+import type { Offer } from '../../types';
+import { Favorites, Locations } from '../../components';
 
-export function FavoritesPage(): JSX.Element {
-  const offers = useAppSelector((state) => state.offers);
+type Props = Readonly<{
+  offers: Offer[];
+}>;
 
+export function FavoritesPage(props: Props): JSX.Element {
   return (
     <div className="page__favorites-container container">
       <Helmet>
@@ -21,7 +23,7 @@ export function FavoritesPage(): JSX.Element {
                 <Locations.Item key={item.name} city={item} />
               ))}
             </div>
-            <Favorites offers={offers} />
+            <Favorites {...props} />
           </li>
         </ul>
       </section>
