+++ src/types/state.ts
@@ -2,8 +2,8 @@ import { store } from '../store';
 import { City, Offer, Review } from '.';
 
 export type State = Readonly<{
+  city: City;
   offers: Offer[];
-  city: null | City;
   reviews: Review[];
 }>;
 export type AppDispatch = typeof store.dispatch;
